* Update KeStartTask, WindowProc and WidgetEventHandler parameters.
This WILL NOT break existing applications, as on i686, long is just int.
However, this will lead to better portability later down the line.

* Update version number to V1.02

+++ include/shell.h
@@ -7,6 +7,6 @@
 #ifndef _SHELL_H
 #define _SHELL_H
 
-void ShellRun(int unused_arg);
+void ShellRun(long unused_arg);
 
 #endif//_SHELL_H
