+++ src/wapp/alttab.c
@@ -61,7 +61,7 @@ void AltTab$CalcWindowIndexOrder()
 		g_windowDrawOrderIndex = 0;
 }
 
-void AltTabProc(Window* pWindow, int messageType, int parm1, int parm2)
+void AltTabProc(Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
