+++ src/wapp/cpanel/terminal.c
@@ -23,7 +23,7 @@ enum
 	NTERM_CHANGED_TEXT,
 };
 
-void CALLBACK CplTerminalWndProc(Window* pWindow, int messageType, int parm1, int parm2)
+void CALLBACK CplTerminalWndProc(Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
