+++ src/wapp/crashrep.c
@@ -67,7 +67,7 @@ CrashWinParms;
 #define CRASH_REP_WIN_WIDTH  (500)
 #define CRASH_REP_WIN_HEIGHT (260)
 
-void CALLBACK CrashReportWndProc( Window* pWindow, int messageType, int parm1, int parm2 )
+void CALLBACK CrashReportWndProc( Window* pWindow, int messageType, long parm1, long parm2)
 {
 	switch (messageType)
 	{
@@ -222,7 +222,7 @@ void CrashReporterProcessCrash(CrashInfo* pCrashInfo)
 	
 }
 
-void CrashReporterTask(UNUSED int arg)
+void CrashReporterTask(UNUSED long arg)
 {
 	CrashInfo* pInfo = NULL;
